Improve & simplify valgrind testing (#278)
* Reduce the dependency on the other constructs of this repo, e.g.,
creation & deallocation of Azure VMs via the existing tooling doesn't
seem so much stable. So, introduce a simple docker image that can allow
us running multiple valgrind tests in a single vm in parallel and
anywhere. Also, we don't need to use `create_cluster.sh` and its friends
to create an Azure VM for valgrind testing anymore because it doesn't
make sense anyways. See the changes in README.md for more details.
* Get rid of valgrind code from fab infra because using this infra for
valgrind testing was both unnecessary and this was heavily complicating
the things. Instead, a simpler infra that we can also use in our local
machines seems much more useful and easier to maintain and understand.
* Also remove fabfile/pg_report_vg_query.patch since the Postgres patch
that we depend on to report the query string that causes the memory
error has been merged into Postgres as of PG16.
* Automatically collect stack traces from valgrind core files that are
generated in case of a process crash.

Note to reviewer - **although the PR touches 10 files, this PR mainly
introduces following three files** and updates the README on the usage
of valgrind testing infra. The other 6 files are only modified to remove
the old infra.
* valgrind/Docker/Dockerfile
The dockerfile to build a simple docker image that can be used to run
valgrind tests.
This has nothing but Postgres & Citus built with the flags that are
needed to properly run valgrind tests and valgrind itself.
* valgrind/Docker/run_valgrind_test.sh
The script that we run within the container to run the valgrind test for
the specified test schedule.
We could set the entrypoint of the container to this script but we keep
it as a separate script for simplicity.
* run.sh
The host side script that we use to build the container and run the
valgrind tests through the container.

I strongly recommend reading [the relevant section of the README]
(https://github.com/citusdata/test-automation/tree/main?tab=readme-ov-file#valgrind)
to better understand the changes.